Upgrade to React Native 0.49.3
