+++ ios/ErnRunner/ViewController.m
@@ -10,74 +10,63 @@
 
 - (void)viewDidLoad {
     [super viewDidLoad];
-    // Do any additional setup after loading the view.
-    
+
     MoviesAPI* moviesApi = [[MoviesAPI alloc] init];
-    
+
     [moviesApi.requests registerGetTopRatedMoviesRequestHandlerWithHandler:^(id  _Nullable data, ElectrodeBridgeResponseCompletionHandler  _Nonnull block) {
-        NSLog(@"Invoking request handler");
         NSMutableArray<Movie *> *movies = [[NSMutableArray alloc] init];
-        [movies addObject:[self createMovie:@"1" title:@"The Shawshank Redemption" releaseYear:@1994 rating:@9.2 imageUrl:@"https://images-na.ssl-images-amazon.com/images/M/MV5BMTQ1ODM2MjY3OV5BMl5BanBnXkFtZTgwMTU2MjEyMDE@._V1_.jpg"]];
-        [movies addObject:[self createMovie:@"2" title:@"The Godfather" releaseYear:@1972 rating:@9.2 imageUrl:@"https://images-na.ssl-images-amazon.com/images/M/MV5BZTRmNjQ1ZDYtNDgzMy00OGE0LWE4N2YtNTkzNWQ5ZDhlNGJmL2ltYWdlL2ltYWdlXkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_SY1000_CR0,0,704,1000_AL_.jpg"]];
-        [movies addObject:[self createMovie:@"3" title:@"The Godfather: Part II" releaseYear:@1974 rating:@9 imageUrl:@"https://images-na.ssl-images-amazon.com/images/M/MV5BMGM0MzkxM2MtYWQ2My00NjYyLThhYmYtMTdkNTMyNmFmYTY1XkEyXkFqcGdeQXVyNjU0OTQ0OTY@._V1_SY1000_CR0,0,701,1000_AL_.jpg"]];
-        [movies addObject:[self createMovie:@"4" title:@"The Dark Knight" releaseYear:@2008 rating:@9 imageUrl:@"https://images-na.ssl-images-amazon.com/images/M/MV5BMTMxNTMwODM0NF5BMl5BanBnXkFtZTcwODAyMTk2Mw@@._V1_SY1000_CR0,0,675,1000_AL_.jpg"]];
-        [movies addObject:[self createMovie:@"5" title:@"12 Angry Men" releaseYear:@1957 rating:@8.9 imageUrl:@"https://images-na.ssl-images-amazon.com/images/M/MV5BMWU4N2FjNzYtNTVkNC00NzQ0LTg0MjAtYTJlMjFhNGUxZDFmXkEyXkFqcGdeQXVyNjc1NTYyMjg@._V1_SY1000_CR0,0,649,1000_AL_.jpg"]];
-        
-        NavigationAPI *navigationAPI = [[NavigationAPI alloc] init];
-        [navigationAPI.requests registerNavigateRequestHandlerWithHandler:^(id  _Nullable data, ElectrodeBridgeResponseCompletionHandler  _Nonnull block) {
-            AppDelegate *appDelegate = (AppDelegate *)[[UIApplication sharedApplication] delegate];
-            
-            NavigateData *navData = (NavigateData *)data;
-            NSMutableDictionary *initialPapyload = [[NSMutableDictionary alloc]init];
-            [initialPapyload setObject:navData.initialPayload forKey:@"payload"];
-            
-            UIViewController *viewController = [[ElectrodeReactNative sharedInstance] miniAppWithName:navData.miniAppName properties:initialPapyload];
-            viewController.view.frame = [UIScreen mainScreen].bounds;
-            
-            UINavigationController *navController = (UINavigationController *) appDelegate.window.rootViewController;
-            [navController pushViewController:viewController animated:NO];
-            
-            block(nil, nil);
-        }];
-        
+
+        [movies addObject:[self createMovie:@"1" title:@"The Shawshank Redemption" releaseYear:@1994 rating:@9.2 imageUrl:@"http://bit.ly/2xZm1Zr"]];
+        [movies addObject:[self createMovie:@"2" title:@"The Godfather" releaseYear:@1972 rating:@9.2 imageUrl:@"http://bit.ly/2wK5TuA"]];
+        [movies addObject:[self createMovie:@"3" title:@"The Godfather: Part II" releaseYear:@1974 rating:@9 imageUrl:@"http://bit.ly/2yysiIA"]];
+        [movies addObject:[self createMovie:@"4" title:@"The Dark Knight" releaseYear:@2008 rating:@9 imageUrl:@"http://bit.ly/2iZPBqw"]];
+        [movies addObject:[self createMovie:@"5" title:@"12 Angry Men" releaseYear:@1957 rating:@8.9 imageUrl:@"http://bit.ly/2xwkt7r"]];
+
         block(movies, nil);
-        
-        
     }];
-    
-    
-    UIViewController *viewController =
-    [[ElectrodeReactNative sharedInstance] miniAppWithName:MainMiniAppName properties:nil];
-    [viewController setTitle:@"Top Movies List"];
-    viewController.view.frame = [UIScreen mainScreen].bounds;
-    [self pushViewController:viewController animated:NO];
+
+     NavigationAPI *navigationAPI = [[NavigationAPI alloc] init];
+    [navigationAPI.requests registerNavigateRequestHandlerWithHandler:^(id  _Nullable data, ElectrodeBridgeResponseCompletionHandler  _Nonnull block) {
+        AppDelegate *appDelegate = (AppDelegate *)[[UIApplication sharedApplication] delegate];
+
+        NavigateData *navData = (NavigateData *)data;
+        NSMutableDictionary *initialPapyload = [[NSMutableDictionary alloc]init];
+        [initialPapyload setObject:navData.initialPayload forKey:@"payload"];
+
+        UIViewController *viewController = [[ElectrodeReactNative sharedInstance] miniAppWithName:navData.miniAppName properties:initialPapyload];
+        viewController.view.frame = [UIScreen mainScreen].bounds;
+        [viewController setTitle:@"MovieDetails MiniApp"];
+
+
+        UINavigationController *navController = (UINavigationController *) appDelegate.window.rootViewController;
+        [navController pushViewController:viewController animated:NO];
+
+        block(nil, nil);
+    }];
+
+
+   UIViewController *viewController =
+   [[ElectrodeReactNative sharedInstance] miniAppWithName:MainMiniAppName properties:nil];
+   [viewController setTitle:@"MovieList MiniApp"];
+   viewController.view.frame = [UIScreen mainScreen].bounds;
+   self.navigationBar.translucent = NO;
+   [self pushViewController:viewController animated:NO];
 }
 
 - (void)didReceiveMemoryWarning {
     [super didReceiveMemoryWarning];
-    // Dispose of any resources that can be recreated.
 }
 
 - (Movie *) createMovie : (NSString*) movieId title: (NSString*) title releaseYear: (NSNumber*) releaseYear rating: (NSNumber*) rating imageUrl: (NSString*) imageUrl {
-    
+
     NSMutableDictionary* movieDict = [[NSMutableDictionary alloc]init];
     [movieDict setObject:movieId forKey:@"id"];
     [movieDict setObject:title forKey:@"title"];
     [movieDict setObject:releaseYear forKey:@"releaseYear"];
     [movieDict setObject:imageUrl forKey:@"imageUrl"];
-    
+
     Movie *movie =[[Movie alloc] initWithDictionary:movieDict];
     return movie;
 }
-
-/*
- #pragma mark - Navigation
- 
- // In a storyboard-based application, you will often want to do a little preparation before navigation
- - (void)prepareForSegue:(UIStoryboardSegue *)segue sender:(id)sender {
- // Get the new view controller using [segue destinationViewController].
- // Pass the selected object to the new view controller.
- }
- */
 
 @end
