+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "movielistminiapp",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "private": false,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "16.0.0-alpha.12",
-    "react-native": "0.48.3",
+    "react": "16.0.0-beta.5",
+    "react-native": "0.49.3",
     "react-native-electrode-bridge": "1.5.9",
     "react-native-ernmovie-api": "0.0.9",
     "react-native-ernnavigation-api": "0.0.3"
